refactor: extract core models and repositories
