+++ src/stable_yield_lab/core/__init__.py
@@ -1,0 +1,19 @@
+"""Core data structures for :mod:`stable_yield_lab`.
+
+This subpackage groups the fundamental models and repositories used across
+the project so they can be shared without importing the entire public
+interface exposed in :mod:`stable_yield_lab.__init__`.
+"""
+
+from __future__ import annotations
+
+from .models import Pool, PoolReturn
+from .repositories import PoolRepository, ReturnRepository
+
+__all__ = [
+    "Pool",
+    "PoolReturn",
+    "PoolRepository",
+    "ReturnRepository",
+]
+
