+++ tests/test_reporting.py
@@ -7,7 +7,8 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from stable_yield_lab import Pool, PoolRepository
+from stable_yield_lab.core.models import Pool
+from stable_yield_lab.core.repositories import PoolRepository
 from stable_yield_lab.reporting import cross_section_report
